requirements.txt
Flask
pytest
requests
numpy
pandas
Flask-SQLAlchemy
psycopg2-binary

httpx
fastapi
python-jose[cryptography]
uvicorn

passlib[bcrypt]


flask-swagger-ui==4.11.1

bcrypt==4.0.1
chardet>=5.2.0
python-multipart>=0.0.9

python-dotenv>=1.0.1

